move model ptb kernel to /transitions
