+++ pyabc/inference/util.py
@@ -9,8 +9,8 @@ from pyabc.acceptor import Acceptor
 from pyabc.distance import Distance
 from pyabc.epsilon import Epsilon
 from pyabc.model import Model
-from pyabc.random_variables import RV, ModelPerturbationKernel, Distribution
-from pyabc.transition import Transition
+from pyabc.random_variables import RV, Distribution
+from pyabc.transition import Transition, ModelPerturbationKernel
 
 from pyabc.pyabc_rand_choice import fast_random_choice
 from pyabc.parameters import Parameter
